Add #warning statement.
Triggers a warning diagnostic with a string literal inside the statement.
Does not participate in the rest of the compiler pipeline after that.

+++ lib/AST/Attr.cpp
@@ -77,7 +77,7 @@ bool DeclAttribute::canAttributeAppearOnDecl(DeclAttrKind DK, const Decl *D) {
 }
 
 bool DeclAttribute::canAttributeAppearOnDeclKind(DeclAttrKind DAK, DeclKind DK) {
-  unsigned Options = getOptions(DAK);
+  unsigned long Options = getOptions(DAK);
   switch (DK) {
 #define DECL(Id, Parent) case DeclKind::Id: return (Options & On##Id) != 0;
 #include "swift/AST/DeclNodes.def"
@@ -479,7 +479,7 @@ void DeclAttribute::print(llvm::raw_ostream &OS) const {
   print(P, PrintOptions());
 }
 
-unsigned DeclAttribute::getOptions(DeclAttrKind DK) {
+uint64_t DeclAttribute::getOptions(DeclAttrKind DK) {
   switch (DK) {
   case DAK_Count:
     llvm_unreachable("getOptions needs a valid attribute");
